requirements.txt
google-generativeai
langdetect
streamlit
python-dotenv